requirements.txt
authlib==1.6.5
cachetools==6.2.0
Flask==3.1.2
python-dotenv==1.1.1
PyMySQL==1.1.1
SQLAlchemy==2.0.35
